# internal/openapi/swagger.yaml
basePath: /api
definitions:
...
  api.LoginRequest:
    properties:
      login:
        type: string
      password:
        type: string
    type: object
...
info:
  contact: {}
  description: HTTP API for the AWG Manager daemon (tunnels, routing, system).
  title: AWG Manager API
  version: "1.0"
paths:
...
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticates with Keenetic credentials; sets HttpOnly session
        cookie awg_session.
      parameters:
      - description: Router login and password
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: success, login
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "503":
          description: Service Unavailable
          schema:
            additionalProperties: true
            type: object
      summary: Login
      tags:
      - auth
...
  /client-routes:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
      security:
      - CookieAuth: []
      summary: List client routes
      tags:
      - client-routes
  /client-routes/create:
    post:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - CookieAuth: []
      summary: Create client route
      tags:
      - client-routes
  /client-routes/delete:
    post:
      parameters:
      - description: Route id
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - CookieAuth: []
      summary: Delete client route
      tags:
      - client-routes
  /client-routes/toggle:
    post:
      consumes:
      - application/json
      parameters:
      - description: Route id
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - CookieAuth: []
      summary: Toggle client route
      tags:
      - client-routes
  /client-routes/update:
    post:
      consumes:
      - application/json
      parameters:
      - description: Route id
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - CookieAuth: []
      summary: Update client route
      tags:
      - client-routes
...
  /routing/client-routes:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
      security:
      - CookieAuth: []
      summary: List client routes
      tags:
      - client-routes
...
securityDefinitions:
  CookieAuth:
    description: Session cookie set by POST /auth/login. Omit for public routes.
    in: cookie
    name: awg_session
    type: apiKey
swagger: "2.0"